4.0.0/stability Moving components to their own directory
